+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v2.9.7
 
 * add binary shift operations to `AtomicNumber`
-
+* add `default` to `read_in_file`